fix: hall_area_sqmの型をintからfloatに変更
Pydanticバリデーションエラーが発生していた問題を修正:
- 面積データに小数点が含まれる可能性があるため（例: 634.8）
- app/dto/gym.py と app/schemas/gym_detail.py の両方を修正

+++ app/dto/gym.py
@@ -128,7 +128,7 @@ class GymDetailDTO(BaseModel):
 
     # Hall-specific fields
     hall_sports: list[str] = Field(default_factory=list, description="対応スポーツ一覧")
-    hall_area_sqm: int | None = Field(default=None, description="面積（平方メートル）")
+    hall_area_sqm: float | None = Field(default=None, description="面積（平方メートル）")
 
     # Field-specific fields
     field_type: str | None = Field(default=None, description="グラウンドタイプ")
